refactor(profiler): unify NPU and CUDA under FLTorchProfilerWrapper
Per review feedback — collapse AscendTorchProfilerWrapper and the
in-worker TorchProfilerWrapper branch into a single FLTorchProfilerWrapper
that picks its backend at construction time. The hardware probe
(torch_npu + torch.npu.is_available()) moves into the profiler module,
keeping WorkerFL._init_profiler agnostic of the backend choice.

- vllm_fl/profiler/wrapper.py: new FLTorchProfilerWrapper with
  _build_npu_backend (torch_npu.profiler.profile, mirroring vllm-ascend's
  _create_profiler) and _build_cuda_backend (vLLM TorchProfilerWrapper).
- vllm_fl/profiler/ascend.py: removed (superseded).
- vllm_fl/profiler/__init__.py: re-export FLTorchProfilerWrapper.
- vllm_fl/worker/worker.py: drop TorchProfilerWrapper import; replace
  30-line probe+branch with single FLTorchProfilerWrapper call.
