+++ vllm_fl/profiler/__init__.py
@@ -1,3 +1,22 @@
 # Copyright (c) 2025 BAAI. All rights reserved.
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+"""FL profiler wrappers.
+
+``FLTorchProfilerWrapper`` is the single entry point — it internally branches
+between Ascend NPU and CUDA backends. See ``wrapper.py`` for details.
+"""
+
+from vllm_fl.profiler.wrapper import FLTorchProfilerWrapper
+
+__all__ = ["FLTorchProfilerWrapper"]
